# Caching and CMLReader.load performance improvements