--- a/cmlreaders/readers/electrodes.py
+++ b/cmlreaders/readers/electrodes.py
@@ -6,8 +6,10 @@ from pandas.io.json import json_normalize
 
 from cmlreaders import exc
 from cmlreaders.base_reader import BaseCMLReader
+from cmlreaders.cache import cache
 
 
+@cache
 class MontageReader(BaseCMLReader):
     """Reads montage files (contacts.json, pairs.json). When loading via
     :meth:`CMLReader.load`, pass ``read_categories=True`` to additionally load