Closes #111.

## New features

- Optional caching per reader type (only in-memory caching is supported now)
- Ability to globally toggle caching for readers that have enabled caching (default is that it is enabled)
- Load speed improvements by storing the results of the localization/montage number lookups

## High level overview

- `BaseCMLReader` children get a new variable `caching` which is either `None` or a string specifying the type of caching to use (only `"memory"` is supported now)
- Reader instances are cached using `functools.lru_cache`
- `BaseCMLReader` keeps track of all instances so that caches can be cleared globally
- Global settings can be configured in the `cmlreaders.cache` module
- Caching is *enabled* by default since it is in-memory-only right now (this will probably change when support is added for disk-based caching)

## Reader types configured to use in-memory caching

- `MontageReader`
- `LocalizationReader`
- `ElectrodeCategoriesReader`
- `EventReader`

## Typical speedups

First load of pairs:

```ipython
In [5]: %time _ = reader.load("pairs")
CPU times: user 47.4 ms, sys: 6.1 ms, total: 53.5 ms
Wall time: 58.9 ms
```

Subsequent reload:

```ipython
In [6]: %time _ = reader.load("pairs")
CPU times: user 19 µs, sys: 1 µs, total: 20 µs
Wall time: 22.9 µs
```

Events:

```ipython
In [7]: %time _ = reader.load("events")
CPU times: user 30.2 ms, sys: 3.28 ms, total: 33.4 ms
Wall time: 39.1 ms

In [8]: %time _ = reader.load("events")
CPU times: user 19 µs, sys: 1 µs, total: 20 µs
Wall time: 23.1 µs
```