@mivade 2018-07-25
Some issues that need to be resolved:

- The cache decorator screws up a bunch of tests and results in not having access to convenience methods like `fromfile`
- A new instance of `BaseCMLReader` subclasses is made every time `CMLReader.load` is called, thus no cached results are ever used